+++ mayan/apps/converter/tests/mixins.py
@@ -1,4 +1,3 @@
-from django.contrib.contenttypes.models import ContentType
 from django.core.files import File
 from django.db.models import Q
 
@@ -6,14 +5,14 @@ from mayan.apps.acls.classes import ModelPermission
 from mayan.apps.permissions.tests.mixins import PermissionTestMixin
 
 from ..classes import Layer
-from ..models import Asset, ObjectLayer
+from ..models import Asset, LayerTransformation
 from ..transformations import BaseTransformation
 
 from .literals import (
     TEST_ASSET_LABEL, TEST_ASSET_LABEL_EDITED, TEST_ASSET_INTERNAL_NAME,
-    TEST_ASSET_PATH, TEST_TRANSFORMATION_CLASS_LABEL,
-    TEST_TRANSFORMATION_CLASS_NAME, TEST_TRANSFORMATION_NAME,
-    TEST_TRANSFORMATION_ARGUMENT, TEST_TRANSFORMATION_ARGUMENT_EDITED
+    TEST_ASSET_PATH, TEST_LAYER_LABEL, TEST_LAYER_ORDER, TEST_LAYER_NAME,
+    TEST_TRANSFORMATION_ARGUMENT, TEST_TRANSFORMATION_ARGUMENT_EDITED,
+    TEST_TRANSFORMATION_LABEL, TEST_TRANSFORMATION_NAME
 )
 
 
@@ -136,74 +135,116 @@ class LayerTestCaseMixin:
 
 
 class LayerTestMixin(PermissionTestMixin):
-    test_layer = Layer(
-        label='Test layer', name='test_layer', order=1000,
-        permissions={}
-    )
+    auto_create_test_layer = True
 
     def setUp(self):
         super().setUp()
-        self._create_test_permission()
 
-        self.test_layer_permission = self.test_permission
-        ModelPermission.register(
-            model=self.test_document._meta.model, permissions=(
-                self.test_permission,
+        if self.auto_create_test_layer:
+            self.test_layer = Layer(
+                label=TEST_LAYER_LABEL, name=TEST_LAYER_NAME,
+                order=TEST_LAYER_ORDER, permissions={}
             )
-        )
 
-        self.test_layer.permissions = {
-            'create': self.test_layer_permission,
-            'delete': self.test_layer_permission,
-            'edit': self.test_layer_permission,
-            'select': self.test_layer_permission,
-            'view': self.test_layer_permission,
-        }
+            self._create_test_permission()
+
+            self.test_layer_permission = self.test_permission
+            ModelPermission.register(
+                model=self.test_document._meta.model, permissions=(
+                    self.test_permission,
+                )
+            )
+
+            self.test_layer.permissions = {
+                'create': self.test_layer_permission,
+                'delete': self.test_layer_permission,
+                'edit': self.test_layer_permission,
+                'select': self.test_layer_permission,
+                'view': self.test_layer_permission,
+            }
+
+    def tearDown(self):
+        if self.auto_create_test_layer:
+            Layer._registry.pop(self.test_layer.name, None)
+
+        super().tearDown()
 
 
 class TransformationTestMixin(LayerTestMixin):
-    def _create_test_transformation(self):
-        content_type = ContentType.objects.get_for_model(
-            model=self.test_document
-        )
-        object_layer, created = ObjectLayer.objects.get_or_create(
-            content_type=content_type, object_id=self.test_document.pk,
-            stored_layer=self.test_layer.stored_layer
+    auto_create_test_transformation_class = True
+
+    def setUp(self):
+        super().setUp()
+        if self.auto_create_test_transformation_class:
+            self._create_test_transformation_class()
+
+        BaseTransformation.register(
+            layer=self.test_layer, transformation=self.TestTransformationClass
         )
 
-        self.test_transformation = object_layer.transformations.create(
-            name=TEST_TRANSFORMATION_NAME,
-            arguments=TEST_TRANSFORMATION_ARGUMENT
+    def _create_test_transformation(self):
+        self.test_transformation = self.test_layer.add_transformation_to(
+            obj=self.test_document,
+            transformation_class=self.TestTransformationClass,
+            arguments=getattr(
+                self, 'test_transformation_arguments',
+                TEST_TRANSFORMATION_ARGUMENT
+            )
         )
 
     def _create_test_transformation_class(self):
         class TestTransformation(BaseTransformation):
             arguments = ('test_argument',)
-            label = TEST_TRANSFORMATION_CLASS_LABEL
-            name = TEST_TRANSFORMATION_CLASS_NAME
+            label = TEST_TRANSFORMATION_LABEL
+            name = TEST_TRANSFORMATION_NAME
 
             def execute_on(self, *args, **kwargs):
                 super().execute_on(*args, **kwargs)
 
                 return self.image
 
-        BaseTransformation.register(
-            layer=self.test_layer, transformation=TestTransformation
-        )
-
         self.TestTransformationClass = TestTransformation
 
 
 class TransformationViewsTestMixin:
-    def _request_transformation_create_view(self):
-        return self.post(
+    def _request_transformation_create_post_view(self):
+        pk_list = list(LayerTransformation.objects.values('pk'))
+
+        response = self.post(
             viewname='converter:transformation_create', kwargs={
                 'app_label': 'documents', 'model_name': 'document',
                 'object_id': self.test_document.pk,
                 'layer_name': self.test_layer.name,
-                'transformation_name': TEST_TRANSFORMATION_NAME,
+                'transformation_name': self.TestTransformationClass.name,
             }, data={
-                'arguments': TEST_TRANSFORMATION_ARGUMENT
+                'arguments': getattr(
+                    self, '.test_transformation_argument',
+                    TEST_TRANSFORMATION_ARGUMENT
+                )
+            }
+        )
+
+        try:
+            self.test_transformation = LayerTransformation.objects.get(
+                ~Q(pk__in=pk_list)
+            )
+        except LayerTransformation.DoesNotExist:
+            self.test_transformation = None
+
+        return response
+
+    def _request_transformation_create_get_view(self):
+        return self.get(
+            viewname='converter:transformation_create', kwargs={
+                'app_label': 'documents', 'model_name': 'document',
+                'object_id': self.test_document.pk,
+                'layer_name': self.test_layer.name,
+                'transformation_name': self.TestTransformationClass.name,
+            }, data={
+                'arguments': getattr(
+                    self, '.test_transformation_argument',
+                    TEST_TRANSFORMATION_ARGUMENT
+                )
             }
         )
 
@@ -221,7 +262,10 @@ class TransformationViewsTestMixin:
                 'layer_name': self.test_layer.name,
                 'transformation_id': self.test_transformation.pk
             }, data={
-                'arguments': TEST_TRANSFORMATION_ARGUMENT_EDITED
+                'arguments': getattr(
+                    self, 'test_transformation_argument_edited',
+                    TEST_TRANSFORMATION_ARGUMENT_EDITED
+                )
             }
         )
 
@@ -250,6 +294,6 @@ class TransformationViewsTestMixin:
                 'object_id': self.test_document.pk,
                 'layer_name': self.test_layer.name
             }, data={
-                'transformation': TEST_TRANSFORMATION_CLASS_NAME
+                'transformation': self.TestTransformationClass.name
             }
         )