+++ mayan/apps/converter/tests/test_transformation_views.py
@@ -9,12 +9,12 @@ class TransformationViewsTestCase(
     TransformationTestMixin, TransformationViewsTestMixin,
     GenericDocumentViewTestCase
 ):
-    def test_transformation_create_view_no_permission(self):
+    def test_transformation_create_post_view_no_permission(self):
         transformation_count = LayerTransformation.objects.count()
 
         self._clear_events()
 
-        response = self._request_transformation_create_view()
+        response = self._request_transformation_create_post_view()
         self.assertEqual(response.status_code, 404)
 
         self.assertEqual(
@@ -24,19 +24,54 @@ class TransformationViewsTestCase(
         events = self._get_test_events()
         self.assertEqual(events.count(), 0)
 
-    def test_transformation_create_view_with_permission(self):
+    def test_transformation_create_post_view_with_permission(self):
         self.grant_access(
-            obj=self.test_document, permission=self.test_permission
+            obj=self.test_document,
+            permission=self.test_layer.permissions['create']
         )
         transformation_count = LayerTransformation.objects.count()
 
         self._clear_events()
 
-        response = self._request_transformation_create_view()
+        response = self._request_transformation_create_post_view()
         self.assertEqual(response.status_code, 302)
 
         self.assertEqual(
             LayerTransformation.objects.count(), transformation_count + 1
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_transformation_create_get_view_no_permission(self):
+        transformation_count = LayerTransformation.objects.count()
+
+        self._clear_events()
+
+        response = self._request_transformation_create_get_view()
+        self.assertEqual(response.status_code, 404)
+
+        self.assertEqual(
+            LayerTransformation.objects.count(), transformation_count
+        )
+
+        events = self._get_test_events()
+        self.assertEqual(events.count(), 0)
+
+    def test_transformation_create_get_view_with_permission(self):
+        self.grant_access(
+            obj=self.test_document,
+            permission=self.test_layer.permissions['create']
+        )
+        transformation_count = LayerTransformation.objects.count()
+
+        self._clear_events()
+
+        response = self._request_transformation_create_get_view()
+        self.assertEqual(response.status_code, 200)
+
+        self.assertEqual(
+            LayerTransformation.objects.count(), transformation_count
         )
 
         events = self._get_test_events()
@@ -61,7 +96,8 @@ class TransformationViewsTestCase(
     def test_transformation_delete_view_with_access(self):
         self._create_test_transformation()
         self.grant_access(
-            obj=self.test_document, permission=self.test_layer_permission
+            obj=self.test_document,
+            permission=self.test_layer.permissions['delete']
         )
         transformation_count = LayerTransformation.objects.count()
 
@@ -97,7 +133,8 @@ class TransformationViewsTestCase(
     def test_transformation_edit_view_with_access(self):
         self._create_test_transformation()
         self.grant_access(
-            obj=self.test_document, permission=self.test_layer_permission
+            obj=self.test_document,
+            permission=self.test_layer.permissions['edit']
         )
         transformation_arguments = self.test_transformation.arguments
 
@@ -135,7 +172,8 @@ class TransformationViewsTestCase(
     def test_transformation_list_view_with_access(self):
         self._create_test_transformation()
         self.grant_access(
-            obj=self.test_document, permission=self.test_permission
+            obj=self.test_document,
+            permission=self.test_layer.permissions['view']
         )
 
         self._clear_events()
@@ -154,7 +192,6 @@ class TransformationViewsTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_transformation_select_get_view_no_permission(self):
-        self._create_test_transformation_class()
         transformation_count = LayerTransformation.objects.count()
 
         self._clear_events()
@@ -170,9 +207,9 @@ class TransformationViewsTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_transformation_select_get_view_with_access(self):
-        self._create_test_transformation_class()
         self.grant_access(
-            obj=self.test_document, permission=self.test_permission
+            obj=self.test_document,
+            permission=self.test_layer.permissions['select']
         )
         transformation_count = LayerTransformation.objects.count()
 
@@ -189,7 +226,6 @@ class TransformationViewsTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_transformation_select_post_view_no_permission(self):
-        self._create_test_transformation_class()
         transformation_count = LayerTransformation.objects.count()
 
         self._clear_events()
@@ -205,9 +241,9 @@ class TransformationViewsTestCase(
         self.assertEqual(events.count(), 0)
 
     def test_transformation_select_post_view_with_access(self):
-        self._create_test_transformation_class()
         self.grant_access(
-            obj=self.test_document, permission=self.test_permission
+            obj=self.test_document,
+            permission=self.test_layer.permissions['select']
         )
         transformation_count = LayerTransformation.objects.count()